添加租户ID的功能 for saas

+++ Persistence/Hzdtf.MySql.Standard/MySqlDapperBase.cs
@@ -14,8 +14,9 @@ namespace Hzdtf.MySql.Standard
     /// MySql Dapper基类
     /// @ 黄振东
     /// </summary>
+    /// <typeparam name="IdT">ID类型</typeparam>
     /// <typeparam name="ModelT">模型类型</typeparam>
-    public abstract partial class MySqlDapperBase<ModelT> : DapperPersistenceBase<ModelT> where ModelT : SimpleInfo
+    public abstract partial class MySqlDapperBase<IdT, ModelT> : DapperPersistenceBase<IdT, ModelT> where ModelT : SimpleInfo<IdT>
     {
         #region 属性与字段
 
@@ -36,7 +37,7 @@ namespace Hzdtf.MySql.Standard
         /// <param name="id">ID</param>
         /// <param name="propertyNames">属性名称集合</param>
         /// <returns>SQL语句</returns>
-        protected override string SelectSql(int id, string[] propertyNames = null) => $"{SelectSql(propertyNames: propertyNames)} {WHERE_ID_EQUAL_PARAM_SQL}";
+        protected override string SelectSql(IdT id, string[] propertyNames = null) => $"{BasicSelectSql(propertyNames: propertyNames)} {WHERE_ID_EQUAL_PARAM_SQL} {GetTenantIdFilterSql(isBeforeAppAnd: true)}";
 
         /// <summary>
         /// 根据ID集合查询模型列表SQL语句
@@ -45,14 +46,14 @@ namespace Hzdtf.MySql.Standard
         /// <param name="parameters">参数</param>
         /// <param name="propertyNames">属性名称集合</param>
         /// <returns>SQL语句</returns>
-        protected override string SelectSql(int[] ids, out DynamicParameters parameters, string[] propertyNames = null) => $"{SelectSql(propertyNames: propertyNames)} WHERE {GetWhereIdsSql(ids, out parameters)}";
+        protected override string SelectSql(IdT[] ids, out DynamicParameters parameters, string[] propertyNames = null) => $"{BasicSelectSql(propertyNames: propertyNames)} WHERE {GetWhereIdsSql(ids, out parameters)} {GetTenantIdFilterSql(isBeforeAppAnd: true)}";
 
         /// <summary>
         /// 根据ID统计模型数SQL语句
         /// </summary>
         /// <param name="id">ID</param>
         /// <returns>SQL语句</returns>
-        protected override string CountSql(int id) => $"{CountSql()} {WHERE_ID_EQUAL_PARAM_SQL}";
+        protected override string CountSql(IdT id) => $"{BasicCountSql()} {WHERE_ID_EQUAL_PARAM_SQL} {GetTenantIdFilterSql(isBeforeAppAnd: true)}";
 
         /// <summary>
         /// 统计模型数SQL语句
@@ -60,6 +61,18 @@ namespace Hzdtf.MySql.Standard
         /// <param name="pfx">前辍</param>
         /// <returns>SQL语句</returns>
         protected override string CountSql(string pfx = null)
+        {
+            string tbAlias = string.IsNullOrWhiteSpace(pfx) ? null : pfx.Replace(".", null);
+
+            return $"{BasicCountSql(pfx)} {GetTenantIdFilterSql(isBeforeAppWhere: true, pfx: tbAlias)}";
+        }
+
+        /// <summary>
+        /// 基本统计模型数SQL语句
+        /// </summary>
+        /// <param name="pfx">前辍</param>
+        /// <returns>SQL语句</returns>
+        protected string BasicCountSql(string pfx = null)
         {
             string tbAlias = string.IsNullOrWhiteSpace(pfx) ? null : pfx.Replace(".", null);
 
@@ -74,6 +87,28 @@ namespace Hzdtf.MySql.Standard
         /// <param name="propertyNames">属性名称集合</param>
         /// <returns>SQL语句</returns>
         protected override string SelectSql(string pfx = null, string appendFieldSqls = null, string[] propertyNames = null)
+        {
+            string tbAlias = null;
+            if (string.IsNullOrWhiteSpace(pfx))
+            {
+                pfx = $"`{Table}`.";
+            }
+            else
+            {
+                tbAlias = pfx.Replace(".", null);
+            }
+
+            return $"{BasicSelectSql(pfx, appendFieldSqls, propertyNames)} {GetTenantIdFilterSql(isBeforeAppWhere: true, pfx: tbAlias)}";
+        }
+
+        /// <summary>
+        /// 基本查询模型列表
+        /// </summary>
+        /// <param name="pfx">前辍</param>
+        /// <param name="appendFieldSqls">追加字段SQL，包含前面的,</param>
+        /// <param name="propertyNames">属性名称集合</param>
+        /// <returns>SQL语句</returns>
+        protected string BasicSelectSql(string pfx = null, string appendFieldSqls = null, string[] propertyNames = null)
         {
             string tbAlias = null;
             if (string.IsNullOrWhiteSpace(pfx))
@@ -102,7 +137,8 @@ namespace Hzdtf.MySql.Standard
             StringBuilder whereSql = MergeWhereSql(filter, out parameters);
             string sortSql = GetSelectPageSortSql(filter, GetSelectSortNamePfx(filter));
 
-            return $"{SelectSql(appendFieldSqls: AppendSelectPageFieldsSql(), propertyNames: propertyNames)} {GetSelectPageJoinSql(parameters, filter)} {whereSql.ToString()} {sortSql} {GetPartPageSql(pageIndex, pageSize)}";
+            return $"{BasicSelectSql(appendFieldSqls: AppendSelectPageFieldsSql(), propertyNames: propertyNames)} " +
+                $"{GetSelectPageJoinSql(parameters, filter)} {whereSql.ToString()}  {GetTenantIdFilterSql(isBeforeAppAnd: true)} {sortSql} {GetPartPageSql(pageIndex, pageSize)}";
         }
 
         /// <summary>
@@ -203,7 +239,37 @@ namespace Hzdtf.MySql.Standard
         protected override string CountByFilterSql(FilterInfo filter, out DynamicParameters parameters)
         {
             StringBuilder whereSql = MergeWhereSql(filter, out parameters);
-            return $"{CountSql()} {GetSelectPageJoinSql(parameters, filter)} {whereSql.ToString()}";
+            return $"{BasicCountSql()} {GetSelectPageJoinSql(parameters, filter)} {whereSql.ToString()} {GetTenantIdFilterSql(isBeforeAppAnd: true)}";
+        }
+
+        /// <summary>
+        /// 获取租户ID筛选SQL
+        /// </summary>
+        /// <param name="isBeforeAppWhere">是否前面追加WHERE</param>
+        /// <param name="isBeforeAppAnd">是否前面追加AND</param>
+        /// <param name="pfx">前辍</param>
+        /// <returns>租户ID筛选SQL</returns>
+        protected virtual string GetTenantIdFilterSql(bool isBeforeAppWhere = false, bool isBeforeAppAnd = false, string pfx = null)
+        {
+            IdT tenantId;
+            if (IsExistsTenantId(out tenantId))
+            {
+                var tenantIdField = GetFieldByProp("TenantId");
+                pfx = string.IsNullOrWhiteSpace(pfx) ? null : pfx + ".";
+                var sql = $"{pfx}`{tenantIdField}`={Identity.GetValueSql(tenantId)}";
+                if (isBeforeAppWhere)
+                {
+                    return $" WHERE {sql}";
+                }
+                else if (isBeforeAppAnd)
+                {
+                    return $" AND {sql}";
+                }
+
+                return sql;
+            }
+
+            return null;
         }
 
         #endregion
@@ -218,8 +284,9 @@ namespace Hzdtf.MySql.Standard
         /// <returns>SQL语句</returns>
         protected override string InsertSql(ModelT model, bool isGetAutoId = false)
         {
-            string[] partSql = CombineInsertSqlByFieldNames(InsertFieldNames());
+            string[] partSql = CombineInsertSqlByFieldNames(WrapInsertFieldNames(model.Id));
             string sql = $"INSERT INTO `{Table}`({partSql[0]}) VALUES({partSql[1]})";
+
             return isGetAutoId ? $"{sql};{GetLastInsertIdSql()}" : sql;
         }
 
@@ -231,7 +298,7 @@ namespace Hzdtf.MySql.Standard
         /// <returns>SQL语句</returns>
         protected override string InsertSql(IList<ModelT> models, out DynamicParameters para)
         {
-            string[] partSql = CombineBatchInsertSqlByFieldNames(InsertFieldNames(), models, out para);
+            string[] partSql = CombineBatchInsertSqlByFieldNames(WrapInsertFieldNames(models[0].Id), models, out para);
             return $"INSERT INTO `{Table}`({partSql[0]}) VALUES{partSql[1]}";
         }
 
@@ -241,14 +308,14 @@ namespace Hzdtf.MySql.Standard
         /// <param name="model">模型</param>
         /// <param name="propertyNames">属性名称集合</param>
         /// <returns>SQL语句</returns>
-        protected override string UpdateByIdSql(ModelT model, string[] propertyNames = null) => $"UPDATE `{Table}` SET {GetUpdateFieldsSql(propertyNames)} {WHERE_ID_EQUAL_PARAM_SQL}";
+        protected override string UpdateByIdSql(ModelT model, string[] propertyNames = null) => $"UPDATE `{Table}` SET {GetUpdateFieldsSql(propertyNames)} {WHERE_ID_EQUAL_PARAM_SQL} {GetTenantIdFilterSql(isBeforeAppAnd: true)}";
 
         /// <summary>
         /// 根据ID删除模型SQL语句
         /// </summary>
         /// <param name="id">ID</param>
         /// <returns>SQL语句</returns>
-        protected override string DeleteByIdSql(int id) => $"{DeleteSql()} {WHERE_ID_EQUAL_PARAM_SQL}";
+        protected override string DeleteByIdSql(IdT id) => $"{BasicDeleteSql()} {WHERE_ID_EQUAL_PARAM_SQL} {GetTenantIdFilterSql(isBeforeAppAnd: true)}";
 
         /// <summary>
         /// 根据ID数组删除模型SQL语句
@@ -256,13 +323,19 @@ namespace Hzdtf.MySql.Standard
         /// <param name="ids">ID数组</param>
         /// <param name="parameters">参数集合</param>
         /// <returns>SQL语句</returns>
-        protected override string DeleteByIdsSql(int[] ids, out DynamicParameters parameters) => $"{DeleteSql()} WHERE {GetWhereIdsSql(ids, out parameters)}";
+        protected override string DeleteByIdsSql(IdT[] ids, out DynamicParameters parameters) => $"{BasicDeleteSql()} WHERE {GetWhereIdsSql(ids, out parameters)} {GetTenantIdFilterSql(isBeforeAppAnd: true)}";
 
         /// <summary>
         /// 删除所有模型SQL语句
         /// </summary>
         /// <returns>SQL语句</returns>
-        protected override string DeleteSql() => $"DELETE FROM {Table}";
+        protected override string DeleteSql() => $"DELETE FROM `{Table}` {GetTenantIdFilterSql(isBeforeAppWhere: true)}";
+
+        /// <summary>
+        /// 基本删除所有模型SQL语句
+        /// </summary>
+        /// <returns>SQL语句</returns>
+        protected string BasicDeleteSql() => $"DELETE FROM `{Table}`";
 
         /// <summary>
         /// 创建数据库连接
@@ -271,6 +344,12 @@ namespace Hzdtf.MySql.Standard
         /// <returns>数据库连接</returns>
         public override IDbConnection CreateDbConnection(string connectionString) => new MySqlConnection(connectionString);
 
+        /// <summary>
+        /// 模型是否包含租户ID
+        /// </summary>
+        /// <returns>模型是否包含租户ID</returns>
+        protected override bool ModelContainerTenantId() => !string.IsNullOrWhiteSpace(GetFieldByProp("TenantId"));
+
         #endregion
 
         /// <summary>
@@ -278,7 +357,14 @@ namespace Hzdtf.MySql.Standard
         /// </summary>
         /// <param name="table">表名</param>
         /// <returns>SQL语句</returns>
-        protected override string DeleteByTableSql(string table) => $"DELETE FROM `{table}`";
+        protected override string DeleteByTableSql(string table) => $"DELETE FROM `{table}` WHERE 1=1 {GetTenantIdFilterSql(isBeforeAppAnd: true)}";
+
+        /// <summary>
+        /// 基本根据表名删除所有模型SQL语句
+        /// </summary>
+        /// <param name="table">表名</param>
+        /// <returns>SQL语句</returns>
+        protected string BasicDeleteByTableSql(string table) => $"DELETE FROM `{table}`";
 
         /// <summary>
         /// 根据表名、外键字段和外键值删除模型SQL语句
@@ -288,7 +374,7 @@ namespace Hzdtf.MySql.Standard
         /// <param name="foreignKeyValues">外键值集合</param>
         /// <param name="parameters">参数</param>
         /// <returns>SQL语句</returns>
-        protected override string DeleteByTableAndForignKeySql(string table, string foreignKeyName, int[] foreignKeyValues, out DynamicParameters parameters)
+        protected override string DeleteByTableAndForignKeySql(string table, string foreignKeyName, IdT[] foreignKeyValues, out DynamicParameters parameters)
         {
             parameters = new DynamicParameters();
             StringBuilder whereSql = new StringBuilder();
@@ -300,7 +386,7 @@ namespace Hzdtf.MySql.Standard
             }
             whereSql.Remove(whereSql.Length - 1, 1);
 
-            return $"{DeleteByTableSql(table)} WHERE `{foreignKeyName}` IN({whereSql.ToString()})";
+            return $"{BasicDeleteByTableSql(table)} WHERE `{foreignKeyName}` IN({whereSql.ToString()}) {GetTenantIdFilterSql(isBeforeAppAnd: true)}";
         }
 
         #endregion
@@ -357,7 +443,7 @@ namespace Hzdtf.MySql.Standard
         /// <param name="prefixTable">表前辍</param>
         /// <param name="idField">ID字段</param>
         /// <returns>ID条件SQL语句</returns>
-        protected string GetWhereIdsSql(int[] ids, out DynamicParameters parameters, string prefixTable = null, string idField = "Id") => GetWhereTypesSql<int>(ids, out parameters, idField, prefixTable);
+        protected string GetWhereIdsSql(IdT[] ids, out DynamicParameters parameters, string prefixTable = null, string idField = "Id") => GetWhereTypesSql<IdT>(ids, out parameters, idField, prefixTable);
 
         /// <summary>
         /// 根据值数组获取条件SQL语句，不包含where
@@ -539,7 +625,7 @@ namespace Hzdtf.MySql.Standard
         /// <returns>修改信息SQL</returns>
         protected string GetModifyInfoSql(ModelT model)
         {
-            if (model is PersonTimeInfo)
+            if (model is PersonTimeInfo<IdT>)
             {
                 string[] modifyProps = new string[] { "ModifierId", "Modifier", "ModifyTime" };
                 StringBuilder sql = new StringBuilder();
@@ -603,6 +689,23 @@ namespace Hzdtf.MySql.Standard
         /// <returns>排序名称</returns>
         protected virtual string ConvertSortName(string sortName) => sortName.FristUpper();
 
+        /// <summary>
+        /// 包装插入字段名集合
+        /// </summary>
+        /// <param name="id">ID</param>
+        /// <returns>插入字段名集合</returns>
+        protected virtual string[] WrapInsertFieldNames(IdT id)
+        {
+            var fields = InsertFieldNames();
+            if (PrimaryKeyIncr(id))
+            {
+                var idField = GetFieldByProp("Id", AllFieldMapProps());
+                return fields.Remove(idField);
+            }
+
+            return fields;
+        }
+
         #endregion
 
         #region 私有方法
@@ -622,6 +725,7 @@ namespace Hzdtf.MySql.Standard
                 fieldBuilder.AppendFormat("`{0}`,", field);
                 valueBuilder.AppendFormat("@{0},", GetPropByField(field));
             }
+
             fieldBuilder.Remove(fieldBuilder.Length - 1, 1);
             valueBuilder.Remove(valueBuilder.Length - 1, 1);
